requirements.txt
numpy
Pillow
scipy
rich
